TA Meeting Update
This is the project I showed to the TA. There is multiple screens that
can be worked on now and I have some kind of page control added in.
Please DO NOT use or expand to the current way the pages are navigated
between. This should use an event listener instead of the shitty way I
used. Either please change this or leave it for me